SearchService/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==1.10.13
python-dotenv==1.0.1
python-jose==3.3.0
